+++ src/extern.h
@@ -3,8 +3,8 @@ extern int g_scene;
 extern long g_DebugDisp;
 extern long g_DebugStageProduction;
 
-extern long gameflag[1024];
-extern long gameflag2[1024];
+extern uint32_t gameflag[1024];
+extern uint32_t gameflag2[1024];
